Fix some bugs that were preventing compilation
Also change height to u8
